requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-APScheduler==1.12.3
redis==4.1.0
celery==5.2.3
python-dotenv==0.19.0
requests==2.26.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
